Updated makefiles

+++ src/simple_soup/models/two_counters/ss_model.h
@@ -3,8 +3,8 @@
 
 #define SS_DYNAMIC
 struct ss_state_s {
-    int a;
-    int b;
+    short a;
+    short b;
 };
 
 #endif //SS_MODEL_H
